Updated indexing system, to perform sentence-based indexing and search

+++ database/process_data.py
@@ -1,7 +1,9 @@
 import os
 import json
 import pandas as pd
+import re
 from database.utils import download_data, clean_text
+
 from sklearn.feature_extraction.text import TfidfVectorizer
 import nltk
 import spacy
@@ -12,12 +14,16 @@ from nltk.stem import WordNetLemmatizer
 from elasticsearch import Elasticsearch
 from elasticsearch.helpers import bulk
 import pickle  # To save TF-IDF indexed data
+
 from database.indexing_system import index_elasticsearch, check_elasticsearch_server, delete_elasticsearch_index, build_sentence_corpus_from_json
+
 from sentence_transformers import models, SentenceTransformer
 from tqdm import tqdm
 from nltk.tokenize import sent_tokenize
 import numpy as np 
+
 from database import export_index
+
 import time
 
 # topics = ['cs.AI', 'cs.CV', 'cs.IR', 'cs.LG', 'cs.CL']
@@ -31,22 +37,7 @@ def load_data(Scibert, max_doc, model=None):
     nltk.download("wordnet")
 
     if Scibert :
-        # # Load the SciBERT model
-        # # model = SentenceTransformer("gsarti/scibert-nli")
-        # word_embedding_model = models.Transformer(
-        #     'gsarti/biobert-nli',
-        #     max_seq_length=128,
-        #     do_lower_case=True
-        # )
-        # pooling_model = models.Pooling(word_embedding_model.get_word_embedding_dimension(),
-        #     pooling_mode_mean_tokens=True,
-        #     pooling_mode_cls_token=False,
-        #     pooling_mode_max_tokens=False
-        # )
-
-        # model = SentenceTransformer(modules=[word_embedding_model, pooling_model])
-
-        cols = ['id', 'title', 'abstract', 'categories', 'prepared_text', 'citations', 'embedding']
+        cols = ['id', 'title', 'abstract', 'categories', 'prepared_text', 'citations', 'paragraph_embeddings']
     else:
         cols = ['id', 'title', 'abstract', 'categories', 'prepared_text', 'citations']
 
@@ -62,21 +53,27 @@ def load_data(Scibert, max_doc, model=None):
             if len(data) > max_doc:
                 break
             doc = json.loads(line)
+            
             if doc['categories'] in topics:
                 doc_id = str(doc['id'])  # Ensure ID is a string
-                processed_title = preprocess_text(doc['title'])
+                processed_title = preprocess_text(doc['title'],mode="title")
                 processed_abstract = preprocess_text(doc['abstract'])
-                prepared_text = processed_title + " \n " + processed_abstract  # Combined text for better indexing
+                prepared_text = [processed_title] + processed_abstract  # Combined text for better indexing
+
                 citations = extract_citations(doc['abstract'])  # Extract citations if needed
                 if Scibert:
-                    #Generates a dense vector representation of text using SciBERT.
+                    # Generates a dense vector representation of text using SciBERT.
+                    paragraph_embeddings = [
+                        {'text': p, 'embedding': model.encode(p).tolist()}
+                        for p in prepared_text if p.strip()
+                    ]
                     # sentences = sent_tokenize(prepared_text)
                     # embedding = np.mean(model.encode(sentences), axis=0)
-                    embedding = model.encode(prepared_text)
-                    data.append([doc_id, processed_title, processed_abstract, doc['categories'], prepared_text, citations, embedding])
+
+                    data.append([doc_id, processed_title, processed_abstract, doc['categories'], prepared_text, citations, paragraph_embeddings])
                 else:
                     data.append([doc_id, processed_title, processed_abstract, doc['categories'], prepared_text, citations])
-
+    
     df_data = pd.DataFrame(data=data, columns=cols)
     print("Number of documents loaded:", df_data.shape[0])
     return df_data
@@ -93,16 +90,35 @@ def tfidf(prepared_text):
     
     return tfidf_vectors
 
-def preprocess_text(text):
+def preprocess_text(text, mode=None):
     """Lowercases, tokenizes, removes stopwords, and lemmatizes text."""
     lemmatizer = WordNetLemmatizer()
     stop_words = set(stopwords.words("english"))
 
-    text = text.lower()  # Lowercasing
-    text = text.translate(str.maketrans("", "", string.punctuation))  # Remove punctuation
-    tokens = word_tokenize(text)  # Tokenization
-    tokens = [lemmatizer.lemmatize(word) for word in tokens if word not in stop_words]  # Lemmatization & Stopword Removal
-    return " ".join(tokens)
+    if mode == "title":
+        text = text.lower()  # Lowercasing
+        text = text.translate(str.maketrans("", "", string.punctuation))  # Remove punctuation
+        tokens = word_tokenize(text)  # Tokenization
+        tokens = [lemmatizer.lemmatize(word) for word in tokens if word not in stop_words]  # Lemmatization & Stopword Removal
+        tokens = " ".join(tokens)
+        return tokens
+    else:
+        result = []
+        tmp = text.split('\n')
+        # Dropping the unwant elements (space)
+        text = [i.strip() for i in tmp if i.strip() != '']
+        # Combine to a single string
+        text = ''.join(text)
+        # Add space after fullstop
+        text = re.sub(r'\.([A-Z])', r'. \1', text)
+        # Seperate sentences
+        sentences = sent_tokenize(text)
+        for sentence in sentences:
+            tokens = word_tokenize(sentence) 
+            tokens = [lemmatizer.lemmatize(word.lower().translate(str.maketrans("", "", string.punctuation))) for word in tokens if word not in stop_words]  # Lemmatization & Stopword Removal
+            tokens = " ".join(tokens)
+            result.append(tokens)
+        return result
 
 def extract_citations(text):
     # Load Spacy NLP model for citation extraction
@@ -138,10 +154,10 @@ def test():
     for doc in documents[:5]:  # Print only the first 5 for readability
         print(f"ID: {doc['_id']}, Title: {doc['_source']['title']}, Abstract: {doc['_source']['abstract']}")
 
-def build_index_system(index_name = "arxiv_index", use_bert=True, max_doc=500):
+def build_index_system(index_name = "arxiv_index", use_bert=True, max_doc=500, remove_index=False):
     if use_bert:
         word_embedding_model = models.Transformer(
-            'gsarti/biobert-nli',
+            'allenai/scibert_scivocab_uncased',
             max_seq_length=128,
             do_lower_case=True
         )
@@ -153,20 +169,23 @@ def build_index_system(index_name = "arxiv_index", use_bert=True, max_doc=500):
 
         model = SentenceTransformer(modules=[word_embedding_model, pooling_model])
 
-    delete_elasticsearch_index(index_name)
-    if not check_elasticsearch_server():
+    if remove_index:
+        delete_elasticsearch_index(index_name)
+
+    if True:
         download_data()
         df_data = load_data(use_bert, max_doc, model)
         # Index into Elasticsearch
         index_elasticsearch(df_data, index_name, use_bert)
-    
+
+
     if use_bert:
         time.sleep(10)
         export_index.export_index()
         build_sentence_corpus_from_json(model)
 
-if __name__=="__main__":
-    index_name = "arxiv_index"
-    use_bert = True
-    max_doc=5000
-    build_index_system(index_name, use_bert, max_doc)
+# if __name__=="__main__":
+#     index_name = "arxiv_index"
+#     use_bert = True
+#     max_doc=2000
+#     build_index_system(index_name, use_bert, max_doc, remove_index=True)
